v1.1 complete

+++ API/RPS-API/models/User.cs
@@ -5,8 +5,8 @@ namespace RPS_API.models
     public class User
     {
         public string Username { get; set; }
-        public int Wins { get; set; }
-        public int TurnsPlayed { get; set; }
+        public double Wins { get; set; }
+        public double TurnsPlayed { get; set; }
         public double WinRatio { get; set; }
 
         public User(string username, int wins, int turnsPlayed)
